refactor componentize bindgen from raw_bindgen
