+++ crates/js-component-bindgen/src/identifier.rs
@@ -1,0 +1,36 @@
+pub fn is_js_identifier(s: &str) -> bool {
+  let mut chars = s.chars();
+  if let Some(char) = chars.next() {
+      if !is_js_identifier_start(char) {
+          return false;
+      }
+  } else {
+      return false;
+  }
+  while let Some(char) = chars.next() {
+      if !is_js_identifier_char(char) {
+          return false;
+      }
+  }
+  return true;
+}
+
+// https://tc39.es/ecma262/#prod-IdentifierStartChar
+// Unicode ID_Start | "$" | "_"
+fn is_js_identifier_start(code: char) -> bool {
+  return match code {
+      'A'..='Z' | 'a'..='z' | '$' | '_' => true,
+      // leaving out non-ascii for now...
+      _ => false,
+  };
+}
+
+// https://tc39.es/ecma262/#prod-IdentifierPartChar
+// Unicode ID_Continue | "$" | U+200C | U+200D
+fn is_js_identifier_char(code: char) -> bool {
+  return match code {
+      '0'..='9' | 'A'..='Z' | 'a'..='z' | '$' | '_' => true,
+      // leaving out non-ascii for now...
+      _ => false,
+  };
+}
